skiller ut saf.hentdokument til en egen restClient for å teste det isolert
