+++ src/main/java/no/nav/familie/integrasjoner/journalpost/JournalpostService.kt
@@ -1,12 +1,14 @@
 package no.nav.familie.integrasjoner.journalpost
 
+import no.nav.familie.integrasjoner.client.rest.SafHentDokumentRestClient
 import no.nav.familie.integrasjoner.client.rest.SafRestClient
 import no.nav.familie.integrasjoner.journalpost.domene.Journalpost
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 
 @Service
-class JournalpostService @Autowired constructor(private val safRestClient: SafRestClient) {
+class JournalpostService @Autowired constructor(private val safRestClient: SafRestClient,
+                                                private val safHentDokumentRestClient: SafHentDokumentRestClient) {
 
     fun hentSaksnummer(journalpostId: String): String? {
         val journalpost = safRestClient.hentJournalpost(journalpostId)
@@ -20,7 +22,7 @@ class JournalpostService @Autowired constructor(private val safRestClient: SafRe
     }
 
     fun hentDokument(journalpostId: String, dokumentInfoId: String, variantFormat: String?): ByteArray {
-        val dokument = safRestClient.hentDokument(journalpostId, dokumentInfoId, variantFormat)
+        val dokument = safHentDokumentRestClient.hentDokument(journalpostId, dokumentInfoId, variantFormat)
         return dokument
     }
 }
